% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_nse.R
\name{cal_nse}
\alias{cal_nse}
\title{Calculate the Nash-Sutcliffe efficiency (NSE)}
\usage{
cal_nse(obs, sim, na.rm = T)
}
\arguments{
\item{obs}{Numeric vector of matrix. Measured values}

\item{sim}{Numeric vector of matrix. Predicted values}
}
\value{
The Nash-Sutcliffe efficiency (NSE)
}
\description{
Calculate the Nash-Sutcliffe efficiency (NSE)
}
\references{
Nash, J. E., and J.V. Sutcliffe (1970): River flow forecasting through conceptual models. 1. a discussion of principles. Journal of Hydrology 10, 282–290.
}
