% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_poly_area.R
\name{cal_poly_area}
\alias{cal_poly_area}
\title{Calculate the area of the polygon with the given vextices}
\usage{
cal_poly_area(x, y)
}
\arguments{
\item{x}{Numeric. An array contains the x coordinates of the vextices.}

\item{y}{Numeric. An array contains the y coordinates of the vextices.}
}
\value{
The area of the polyon.
}
\description{
Calculate the area of the polygon with the given vextices
}
\examples{
\dontrun{
x <- c(0, 0, 1, 1)
y <- c(1, 0, 0, 1)
cal_poly_area(x, y)
}
}
