% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_plot.R
\name{calibration_plot}
\alias{calibration_plot}
\title{Generate calibration plots for EFDC exported data.}
\usage{
calibration_plot(fname, ts_num, ...)
}
\arguments{
\item{fname}{Character. The name of the \code{.dat} file exported by EFDC.}

\item{ts_num}{Number. The number of the time series to plot.}

\item{...}{Not used.}
}
\description{
Generate calibration plots for EFDC exported data.
}
