% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse.R
\name{check_dim_}
\alias{check_dim_}
\title{Check if all the inputs are same in dimensions}
\usage{
check_dim_(...)
}
\arguments{
\item{...}{Objects to be checked}
}
\description{
Check if all the inputs are same in dimensions
}
