% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_water_level.R
\name{col_format_water_level}
\alias{col_format_water_level}
\title{Tidy one column of water level data (Just for batch processing).}
\usage{
col_format_water_level(x)
}
\arguments{
\item{x}{Character. One column of water level data.}
}
\description{
Tidy one column of water level data (Just for batch processing).
}
