% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_to_wq.R
\name{dt_to_wq}
\alias{dt_to_wq}
\title{Convert csv data files into wq formats}
\usage{
dt_to_wq(measure.vars, src.dt, path, start.date, end.date, interval = "1 day")
}
\arguments{
\item{measure.vars}{vector of measure variables. Can be integer (corresponding measure column numbers)
or character (measure column names) vector. If missing, all non-id columns will be assigned to it.}

\item{src.dt}{A \code{data.table} object. If a \code{data.frame} object is provided, it will be
converted to a \code{data.table} object using \link[data.table]{setDT}.}

\item{path}{Character. A character indicating the path of the \code{.wq} files to be saved.}

\item{start.date}{Character. A character of desired date ('2000-01-01')}

\item{end.date}{Character. A character of desired date ('2000-01-01')}

\item{interval}{Character. Increment of the sequence. See \link[base]{seq.Date} for details.}
}
\description{
Convert csv data files into wq formats
}
