% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efdc_nc_get_grid.R
\name{efdc_nc_get_grid_dt}
\alias{efdc_nc_get_grid_dt}
\title{Export grid information to a \code{data.table} from the \code{.nc} file generated by EE.}
\usage{
efdc_nc_get_grid_dt(filename)
}
\arguments{
\item{filename}{Character. The name of the \code{.nc} file generated by EE.}
}
\value{
A \code{data.table} contains grid coordinates and elevation.
}
\description{
Export grid information to a \code{data.table} from the \code{.nc} file generated by EE.
}
