% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_rfggrid.R
\name{format_rfggrid}
\alias{format_rfggrid}
\title{Format Delft3D RFGGRID (.grd) file.}
\usage{
format_rfggrid(src_grd, dst_grd,
  na_pattern = "-9.99999000000000024E\\\\+02",
  na_replacement = " 0.00000000000000000E+00", coord_sys = "Cartesian")
}
\arguments{
\item{src_grd}{Character. Full name of the Delft3D RFGGRID (.grd) file to be formatted.}

\item{dst_grd}{Character. Full name of the formatted Delft3D RFGGRID (.grd) file.}

\item{na_pattern}{Character. The MISSING VALUE in the src_grd.}

\item{na_replacement}{Character. A character vector of replacements.}

\item{coord_sys}{Character. Coordinate system in the grd.}
}
\description{
Format Delft3D RFGGRID (.grd) file.
}
\examples{
\dontrun{
f <- system.file("extdata/poyang_lake.grd", package="efdcr")
format_rfggrid(src_grd = f, dst_grd = tempfile(fileext = ".grd"))
}
}
