% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_velocity.R
\name{geom_velocity}
\alias{geom_velocity}
\title{Geom for plotting velocity vectors.}
\usage{
geom_velocity(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  scale = 1,
  arrow_scale = 0.5,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2]{aes}}
or \code{\link[ggplot2]{aes_}}.
If specified and \code{inherit.aes = TRUE} (the default),
it is combined with the default mapping at the top level of the plot.
You must supply mapping if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three options:

If \code{NULL}, the default, the data is inherited from the plot data as specified in the call to \code{\link[ggplot2]{ggplot}}.

A \code{data.frame}, or other object, will override the plot data.
All objects will be fortified to produce a data frame.
See \code{\link[ggplot2]{fortify}} for which variables will be created.}

\item{stat}{Use to override the default connection}

\item{position}{Position adjustment, either as a string, or the result of a call to a position adjustment function.}

\item{na.rm}{If \code{False}, the default, missing values are removed with a warning.
If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE}never includes,
and \code{TRUE} always includes}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics, rather than combining with them.
This is most useful for helper functions that define both data and aesthetics and shouldn't inherit
behaviour from the default plot specification.}

\item{scale}{Multiplicative scaling factor.}

\item{...}{other arguments passed on to \code{\link[ggplot2]{layer}}.}
}
\description{
Geom for plotting velocity vectors.
}
