% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_velocity.R
\docType{data}
\name{GeomVelocity}
\alias{GeomVelocity}
\title{Geom Velocity}
\format{An object of class \code{GeomVelocity} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.}
\usage{
GeomVelocity
}
\description{
Geom Velocity
}
\keyword{datasets}
