% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_efdc_grids_as_sf.R
\name{get_efdc_grids_as_sf}
\alias{get_efdc_grids_as_sf}
\title{Get EFDC grids from lxly or dxly file and convert it to the sf object}
\usage{
get_efdc_grids_as_sf(flxly, fdxdy, fnc)
}
\arguments{
\item{flxly}{Character. Name of the lxly.inp}

\item{fdxdy}{Character. Name of the dxdy.inp}

\item{fnc}{Character. Name of the \code{.nc} file.}
}
\value{
an object of class \code{sf}, which is a classed list-column with simple feature geometries
}
\description{
Get EFDC grids from lxly or dxly file and convert it to the sf object
}
