% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_efdc_nc_dt.R
\name{get_efdc_nc_dt}
\alias{get_efdc_nc_dt}
\title{Export grid information to a \code{data.table} from the \code{.nc} file generated by EE.}
\usage{
get_efdc_nc_dt(filename, varname, wetdepth = 0.15)
}
\arguments{
\item{filename}{Character. The name of the \code{.nc} file generated by EE.}

\item{varname}{Character. The name of the variable to read.}

\item{wetdepth}{Numeric. The minimum depth of the wet cell.}
}
\value{
A \code{data.table} contains grid coordinates and elevation.
}
\description{
Export grid information to a \code{data.table} from the \code{.nc} file generated by EE.
}
