% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_efdc_nc_dt.R
\name{has_variable}
\alias{has_variable}
\title{List the variable names in the NetCDF file.}
\usage{
has_variable(filename)
}
\arguments{
\item{filename}{Character. The name of the \code{.nc} file generated by EE.}
}
\description{
List the variable names in the NetCDF file.
}
