% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse.R
\name{nse}
\alias{nse}
\title{Calculate the Nash-Sutcliffe efficiency (NSE)}
\usage{
nse(obs, sim, na.rm = T)
}
\arguments{
\item{obs}{Numeric vector of matrix. Measured values}

\item{sim}{Numeric vector of matrix. Predicted values}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
The Nash-Sutcliffe efficiency (NSE)
}
\description{
Calculate the Nash-Sutcliffe efficiency (NSE)
}
\references{
Nash, J. E., and J.V. Sutcliffe (1970): River flow forecasting through conceptual models. 1. a discussion of principles. Journal of Hydrology 10, 282–290.
}
