% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs
}
\description{
Pipe an object forward into a function or call expression. 
See \code{\link[magrittr]{\%>\%}} for more details.
}
\keyword{internal}
