% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_efdc_nc_dt.R
\name{plot_efdc_nc_var}
\alias{plot_efdc_nc_var}
\title{Plot EFDC nc data}
\usage{
plot_efdc_nc_var(
  fname,
  var_name = names(nc_open(fname)$var),
  start_date,
  time_unit = "months",
  by_f = lubridate::month
)
}
\arguments{
\item{fname}{Character. The file name of the \code{.nc} file.}

\item{var_name}{Character. The variable name to be plotted.}

\item{start_date}{Character. The start date of the simulation.}

\item{time_unit}{Character. A string that specifies the type of units that num refers to.}

\item{by_f}{function. A function to apply to the date.}
}
\description{
Plot EFDC nc data
}
