% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_efdc_ts.R
\name{plot_efdc_ts}
\alias{plot_efdc_ts}
\title{Generate calibration plots for EFDC exported data.}
\usage{
plot_efdc_ts(fname, ts_num, draw = T, obs_col = "red",
  mod_col = "black", y_lab = NULL, free_scale = T, ...)
}
\arguments{
\item{fname}{Character. The name of the \code{.dat} file exported by EFDC.}

\item{ts_num}{Number. The number of the time series to plot.}

\item{draw}{Logical. If \code{TRUE}, draw the figure.}

\item{obs_col}{Character. The colour of the observed data.}

\item{mod_col}{Character. The colour of the modeled data.}

\item{y_lab}{Character. Label of the Y axis.}

\item{free_scale}{Logical. \code{TRUE} set free scale when facetting.}

\item{...}{Not used.}
}
\description{
Generate calibration plots for EFDC exported data.
}
