% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_area.R
\name{poly_area}
\alias{poly_area}
\title{Calculate the area of the polygon with the given vextices}
\usage{
poly_area(x, y)
}
\arguments{
\item{x}{Numeric. An array contains the x coordinates of the vextices.}

\item{y}{Numeric. An array contains the y coordinates of the vextices.}
}
\value{
The area of the polyon.
}
\description{
Calculate the area of the polygon with the given vextices
}
