% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shp_to_ldb.R
\name{shp_to_ldb}
\alias{shp_to_ldb}
\title{Convert ESRI shapefile into text format.}
\usage{
shp_to_ldb(src_shp, dst_ldb)
}
\arguments{
\item{src_shp}{Character. Full name of the shapfile to be converted.}

\item{dst_ldb}{Character. Full name of the txt file to write.}
}
\value{
Either a list of \code{NULL} or error messages.
}
\description{
Convert ESRI shapefile into text format.
}
