% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_water_level.R
\name{tidy_water_level}
\alias{tidy_water_level}
\title{Tidy the observed water level data}
\usage{
tidy_water_level(x)
}
\arguments{
\item{x}{Character. Full name the water levle file to be handled.}
}
\description{
Tidy the observed water level data
}
