% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_water_level2.R
\name{tidy_water_level2}
\alias{tidy_water_level2}
\title{Tidy the water level data}
\usage{
tidy_water_level2(x)
}
\arguments{
\item{x}{Vector of water level}
}
\value{
Vector of formated water level
}
\description{
Tidy the water level data
}
