% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_wq_params.R
\name{update_wq_params_value}
\alias{update_wq_params_value}
\alias{get_wq_params}
\alias{get_wq_params_value}
\title{Update the water quality parameters in \code{wq3dwc.inp}.}
\usage{
update_wq_params_value(param, new_val, inp)

get_wq_params(card, inp)

get_wq_params_value(param, inp)
}
\arguments{
\item{param}{the name of the parameter to be updated.}

\item{new_val}{the new value of the target water quality parameter.}

\item{inp}{the content of the \code{wq3dwc.inp} file.}

\item{card}{the CARD which the parameter belongs to.}

\item{path}{the path of the updated \code{wq3dwc.inp} files.}
}
\description{
Update the water quality parameters in \code{wq3dwc.inp}.

Get the available water quality parameters.

Get the value of the selected water quality parameter.
}
