% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shp_to_ldb.R
\name{write_ldb}
\alias{write_ldb}
\title{Function for write .ldb files.}
\usage{
write_ldb(x, ldb_file)
}
\arguments{
\item{x}{the data.frame to write.}

\item{ldb_file}{the .ldb file to write with.}
}
\description{
Function for write .ldb files.
}
