% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_to_wq.R
\name{write_wq}
\alias{write_wq}
\title{Write .wq files}
\usage{
write_wq(measure.var_, src.dt_, path_)
}
\arguments{
\item{measure.var_}{vector of measure variables.}

\item{src.dt_}{A \code{data.table} object.}

\item{path_}{Character. A character indicating the path of the \code{.wq} files to be saved.}
}
\description{
Write .wq files
}
