% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_efdc_nc_dt.R
\name{get_efdc_nc_dt}
\alias{get_efdc_nc_dt}
\alias{has_variable}
\alias{get_efdc_nc_var}
\alias{get_efdc_nc_coordinates}
\title{Export grid information to a \code{data.table} from the \code{.nc} file generated by EE.}
\usage{
get_efdc_nc_dt(..., var_name, wet_depth = 0.15, with_coord = T,
  remove_dry = T, verbose = T)

has_variable(filename)

get_efdc_nc_var(..., var_name, wet_depth = 0.15, verbose = T,
  with_coord = T, remove_dry = T)

get_efdc_nc_coordinates(nc, simplify = T)
}
\arguments{
\item{...}{Characters. The names of the \code{.nc} files generated by EE (maximum 2 at once).}

\item{var_name}{Character. The name of the variable to read.}

\item{wet_depth}{Numeric. The minimum depth of the wet cell.}

\item{with_coord}{Logical. Should the data merge with the coordinates (for plot)?}

\item{remove_dry}{Logical. Should the dry cell be removed?}

\item{verbose}{Logical. Should print the processing message?}

\item{filename}{Character. The name of the \code{.nc} file generated by EE}

\item{nc}{NetCDF file connection generated by \code{\link[ncdf4]{nc_open}}.}

\item{simplify}{Logical. If \code{TRUE}, delete unused columns.}
}
\value{
A \code{data.table} contains grid coordinates and elevation.
}
\description{
Export grid information to a \code{data.table} from the \code{.nc} file generated by EE.

List the variable names in the NetCDF file.

Read data from EE generated NetCDF files.

Read coordinates from EE generated NetCDF files.
}
