% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_efdc_nc_dt.R
\name{get_efdc_nc_dt}
\alias{get_efdc_nc_dt}
\alias{melt_nc}
\alias{has_variable}
\alias{get_efdc_nc_var}
\alias{get_efdc_nc_coordinates}
\title{Export grid information to a \code{data.table} from the \code{.nc} file generated by EE.}
\usage{
get_efdc_nc_dt(
  ...,
  var_name,
  wet_depth = 0.15,
  with_coord = T,
  remove_dry = T,
  verbose = T
)

melt_nc(ncarray, var = "value", var_name, nc, na.rm = F)

has_variable(filename)

get_efdc_nc_var(
  ...,
  var_name,
  wet_depth = 0.15,
  verbose = T,
  with_coord = T,
  remove_dry = T
)

get_efdc_nc_coordinates(nc)
}
\arguments{
\item{...}{Characters. The names of the \code{.nc} files generated by EE (maximum 2 at once).}

\item{var_name}{Character. The name of the variable to read.}

\item{wet_depth}{Numeric. The minimum depth of the wet cell.}

\item{with_coord}{Logical. Should the data merge with the coordinates (for plot)?}

\item{remove_dry}{Logical. Should the dry cell be removed?}

\item{verbose}{Logical. Should print the processing message?}

\item{ncarray}{Array. The 3D array read by \code{\link[ncdf4]{ncvar_get}}}

\item{var}{Character. The name of the variable to read.}

\item{nc}{NetCDF file connection generated by \code{\link[ncdf4]{nc_open}}.}

\item{na.rm}{Logical. A logical value indicating whether NA values should be stripped before the computation proceeds.}

\item{filename}{Character. The name of the \code{.nc} file generated by EE}

\item{simplify}{Logical. If \code{TRUE}, delete unused columns.}
}
\value{
A \code{data.table} contains grid coordinates and elevation.
}
\description{
Export grid information to a \code{data.table} from the \code{.nc} file generated by EE.

Melt nc array files

List the variable names in the NetCDF file.

Read data from EE generated NetCDF files.

Read coordinates from EE generated NetCDF files.
}
